Remove /Meetings/Guides and move snippets one level up
