+++ DotNetCliCodeSnippets/DotnetCliCodeSnippets.csproj
@@ -19,4 +19,6 @@
     </None>
   </ItemGroup>
 
+
+
 </Project>
